movie_booking/movie_booking/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.22.0
sqlalchemy>=1.4
psycopg2-binary
bcrypt>=3.2.0,<4.0.0
python-dotenv
aiosmtplib
email-validator
python-jose
redis>=5.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-socketio[asyncio]>=5.11.0
requests>=2.31.0
httpx>=0.24.0
razorpay>=1.3.0
qrcode[pil]>=7.4.2
reportlab>=4.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
